+++ prep/src/cmd/format.rs
@@ -1,34 +1,42 @@
 // Copyright 2026 the Prep Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use anyhow::{Context, ensure};
+use anyhow::{Context, Result, ensure};
+use semver::VersionReq;
 
 use crate::session::Session;
-use crate::tools::cargo::{Cargo, CargoDeps};
+use crate::tools::cargo::CargoDeps;
+use crate::tools::rustfmt::{Rustfmt, RustfmtDeps};
 use crate::ui;
 
 /// Format the workspace.
 ///
 /// In `strict` mode Cargo version is locked.
-pub fn run(session: &mut Session, strict: bool, check: bool) -> anyhow::Result<()> {
-    let mut cmd = if strict {
+pub fn run(session: &mut Session, strict: bool, check: bool) -> Result<()> {
+    let rust_components = vec!["rustfmt".into()];
+    let rustfmt = if strict {
         let tools_cfg = session.config().tools();
         let rustup_ver_req = tools_cfg.rustup().clone();
-        let ver_req = tools_cfg.rust().clone();
+        let cargo_ver_req = tools_cfg.rust().clone();
         let toolset = session.toolset();
-        let deps = CargoDeps::new(rustup_ver_req);
-        toolset.get::<Cargo>(&deps, &ver_req)?
+        let cargo_deps = CargoDeps::new(rustup_ver_req, rust_components);
+        let deps = RustfmtDeps::new(cargo_deps, cargo_ver_req);
+        let ver_req = VersionReq::parse("=1.8.0-stable")?; // TODO: Replace with a custom 'Any' (NOT None!)
+        toolset.get::<Rustfmt>(&deps, &ver_req)?
     } else {
         let toolset = session.toolset();
-        let deps = CargoDeps::new(None);
-        toolset.get::<Cargo>(&deps, None)?
+        let cargo_deps = CargoDeps::new(None, rust_components);
+        let deps = RustfmtDeps::new(cargo_deps, None);
+        toolset.get::<Rustfmt>(&deps, None)?
     };
-    let mut cmd = cmd.current_dir(session.root_dir()).arg("fmt").arg("--all");
+
+    let mut cmd = rustfmt.cmd();
+    cmd.arg("--all");
     if check {
-        cmd = cmd.arg("--check");
+        cmd.arg("--check");
     }
 
-    ui::print_cmd(cmd);
+    ui::print_cmd(&cmd);
 
     let status = cmd.status().context("failed to run cargo fmt")?;
     ensure!(status.success(), "cargo fmt failed: {status}");