+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `tools` command for tool management. ([#27] by [@xStrom])
 * Automatic installation of tools from source via `cargo`.
   Now implemented for `ripgrep`. ([#28] by [@xStrom])
+* Automatic Rustfmt and Clippy installation via `rustup`. ([#29] by [@xStrom])
 * `--strict` option to `clippy`, `copyright`, and `format` commands to use locked tool versions. ([#27] by [@xStrom])
 
 ### Changed
@@ -46,6 +47,7 @@
 [#24]: https://github.com/Nevermore/prep/pull/24
 [#27]: https://github.com/Nevermore/prep/pull/27
 [#28]: https://github.com/Nevermore/prep/pull/28
+[#29]: https://github.com/Nevermore/prep/pull/29
 
 [Unreleased]: https://github.com/Nevermore/prep/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/Nevermore/prep/compare/v0.1.0...v0.2.0